+++ src/index.d.ts
@@ -1,0 +1,19 @@
+
+
+declare module 'httpsnippet' {
+  export default class HttpSnippet {
+    constructor(obj: any);
+    convert(target: string, client: string): string;
+  }
+  export function availableTargets(): Array<{
+    title: string;
+    key: string;
+    extname: string;
+    clients: Array<{
+      title: string;
+      key: string;
+      link?: string;
+      description: string;
+    }>
+  }>;
+}