+++ src/provider/requestCommandsController.ts
@@ -6,7 +6,7 @@ import { extension } from 'mime-types';
 import { promises as fs } from 'fs';
 import { httpDocumentSelector } from '../config';
 import { file } from 'tmp-promise';
-import { toMarkdown } from '../utils';
+import { getHttpRegionFromLine, toMarkdown } from '../utils';
 
 interface CommandData{
   httpRegion: HttpRegion;
@@ -118,7 +118,7 @@ export class RequestCommandsController implements vscode.CodeLensProvider {
 
   @errorHandler()
   async send(document?: vscode.TextDocument, line?: number) {
-    this.httpRegionSendContext = await this.getCurrentHttpRegionSendContext(document, line);
+    this.httpRegionSendContext = await getHttpRegionFromLine(document, line);
     await this.sendRequest(this.httpRegionSendContext);
   }
 
@@ -183,7 +183,7 @@ export class RequestCommandsController implements vscode.CodeLensProvider {
 
   @errorHandler()
   async show(document?: vscode.TextDocument, line?: number) {
-    const parsedDocument = await this.getCurrentHttpRegionSendContext(document, line);
+    const parsedDocument = await getHttpRegionFromLine(document, line);
     if (parsedDocument) {
       await httpYacApi.show(parsedDocument.httpRegion, parsedDocument.httpFile);
     }
@@ -196,7 +196,7 @@ export class RequestCommandsController implements vscode.CodeLensProvider {
       if (this.isHttpRegion(document)) {
         httpRegion = document;
       } else {
-        const parsedDocument = await this.getCurrentHttpRegionSendContext(document, line);
+        const parsedDocument = await getHttpRegionFromLine(document, line);
         if (parsedDocument) {
           httpRegion = parsedDocument.httpRegion;
         }
@@ -217,7 +217,7 @@ export class RequestCommandsController implements vscode.CodeLensProvider {
 
   @errorHandler()
   async save(document?: vscode.TextDocument, line?: number) {
-    const parsedDocument = await this.getCurrentHttpRegionSendContext(document, line);
+    const parsedDocument = await getHttpRegionFromLine(document, line);
     if (parsedDocument && parsedDocument.httpRegion.response) {
       const ext = parsedDocument.httpRegion.metaData.extension || extension(parsedDocument.httpRegion.response.contentType?.contentType || 'application/octet-stream');
       const filters: Record<string, Array<string>> = {};
@@ -240,22 +240,7 @@ export class RequestCommandsController implements vscode.CodeLensProvider {
     await vscode.window.showTextDocument(document);
   }
 
-  private async getCurrentHttpRegionSendContext(doc: vscode.TextDocument  | undefined, line: number | undefined) {
-    const document = doc || vscode.window.activeTextEditor?.document;
-    if (document) {
-      const httpFile = await httpFileStore.getOrCreate(document.fileName, () => Promise.resolve(document.getText()), document.version);
-      if (httpFile) {
-        const currentLine = line ?? vscode.window.activeTextEditor?.selection.active.line;
-        if (currentLine !== undefined) {
-          const httpRegion = httpFile.httpRegions.find(obj => obj.symbol.startLine <= currentLine && currentLine <= obj.symbol.endLine);
-          if (httpRegion) {
-            return { httpRegion, httpFile, httpClient: httpYacApi.httpClient };
-          }
-        }
-      }
-    }
-    return undefined;
-  }
+
 
   toString() {
     return 'requestCommandsController';